♻️ refactor : api 폴더 리팩토링
