+++ src/entities/couple_connect/index.ts
@@ -1,0 +1,4 @@
+export * from "./service";
+  export * from "./queries";
+  export * from "./types";
+
