+++ src/entities/couple_connect/queries.ts
