+++ src/entities/couple_connect/service.ts
