+++ src/entities/couple_nickname/index.ts
@@ -1,0 +1,5 @@
+export * from "./service";
+export * from "./queries";
+export * from "./types";
+
+
