+++ src/entities/init_setting/index.ts
@@ -1,0 +1,7 @@
+export * from "./service";
+export * from "./queries";
+export * from "./types";
+export * from "./mutation";
+
+
+
