+++ src/entities/init_setting/queries.ts
@@ -1,0 +1,14 @@
+import { useQuery } from "@tanstack/react-query";
+import { initSettingQueries } from "./service";
+import { GetCoupleInfoResponse } from "./types";
+
+export const useInitSettingQueries = () => {
+  const getCoupleInfo = useQuery<GetCoupleInfoResponse>({
+    queryKey: ["getCoupleInfo"],
+    queryFn: () => initSettingQueries.getCoupleInfo(),
+  });
+
+  return {
+    getCoupleInfo,
+  };
+};
