+++ src/entities/init_setting/service.ts
@@ -1,7 +1,7 @@
 import { RegisterAnniversaryRequest, RegisterAnniversaryResponse, GetCoupleInfoResponse } from "./types";
 import { api } from "../axios/instance";
 
-export const maonboardingQueries = {
+export const initSettingQueries = {
   registerAnniversary: async (data: RegisterAnniversaryRequest): Promise<RegisterAnniversaryResponse> => {
     const response = await api.post<RegisterAnniversaryResponse>("/couple/anniversary", data);
     return response.data;
