+++ src/entities/init_setting/types.ts
