+++ src/entities/main_anniversary/types.ts
