+++ src/entities/main_letter/types.ts
