+++ src/entities/main_status/types.ts
