+++ src/features/mypage/ui/index.ts
@@ -1,4 +1,4 @@
-export * from "./MyHeader";
+export * from "./MyPageHeader";
 export * from "./EditHeader";
 export * from "./LogoutModal";
 export * from "./DisconnectModal";
